now ft. an actual card controller
