+++ src/main/java/com/revature/quizzard/controllers/CardController.java
@@ -1,0 +1,21 @@
+package com.revature.quizzard.controllers;
+
+import com.revature.quizzard.services.CardService;
+import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/card")
+@AllArgsConstructor(onConstructor = @__(@Autowired))
+public class CardController {
+
+    private final CardService cardService;
+
+    @PostMapping("/favorite")
+    @ResponseStatus(HttpStatus.OK)
+    public void toggleFavoriteCard() {
+
+    }
+}
