+++ src/main/java/com/revature/quizzard/controllers/TestController.java
@@ -3,6 +3,7 @@ package com.revature.quizzard.controllers;
 import com.revature.quizzard.dtos.responsemodel.AccountResponseDTO;
 import com.revature.quizzard.models.user.AccountEntity;
 import com.revature.quizzard.repositories.AccountRepository;
+import com.revature.quizzard.services.CardService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,23 +11,18 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.revature.quizzard.dtos.requestmodels.CardFavoriteDTO;
 
-import com.revature.quizzard.services.AccountService;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 @RestController
-
 @AllArgsConstructor(onConstructor = @__(@Autowired))
 public class TestController {
 
-    private AccountRepository accountRepository;
-    private AccountService accountService;
-
+    private final AccountRepository accountRepository;
+    private final CardService cardService;
 
     @GetMapping("/test")
     public void securityHealthStatus(HttpServletRequest request, HttpServletResponse response){
@@ -49,11 +45,8 @@ public class TestController {
 
     @PostMapping("/favorite/card")
     @ResponseStatus(HttpStatus.OK)
-    public void addFavoriteCard(@RequestBody CardFavoriteDTO dto) {
-        accountService.addFavoriteCard(dto);
+    public void toggleFavoriteCard(@RequestBody CardFavoriteDTO dto) {
+        cardService.addFavoriteCard(dto);
     }
-
-
-
 
 }
