+++ src/main/test/com/revature/quizzard/services/AccountServiceTest.java
@@ -26,18 +26,11 @@ import static org.junit.Assert.*;
 
 public class AccountServiceTest {
 
-
     @InjectMocks
     private AccountService sut;
 
     @Mock
     private AccountRepository mockAccountRepository;
-
-    @Mock
-    private CardRepository mockCardRepository;
-
-    @Mock
-    private AccountCardRepository mockAccountCardRepository;
 
     @Before
     public void setUpTest() {
@@ -46,66 +39,7 @@ public class AccountServiceTest {
 
     @After
     public void tearDownTest() {
-        mockAccountCardRepository = null;
-        mockCardRepository = null;
         mockAccountRepository = null;
     }
 
-    @Test
-    public void when_addFavoriteCardPresent_thenreturnTrue() {
-        AccountEntity accountEntity = new AccountEntity();
-        accountEntity.setId(1);
-        accountEntity.setUsername("username");
-        accountEntity.setPassword("password");
-
-        CardEntity cardEntity = new CardEntity();
-        cardEntity.setQuestion("question");
-        cardEntity.setAnswer("answer");
-        cardEntity.setId(1);
-
-        when(mockAccountRepository.findById(1)).thenReturn(Optional.of(accountEntity));
-        when(mockCardRepository.findById(1)).thenReturn(Optional.of(cardEntity));
-
-        CardFavoriteDTO dto = new CardFavoriteDTO(1, 1, true);
-
-        assertTrue(sut.addFavoriteCard(dto));
-    }
-
-    @Test(expected = InvalidRequestException.class)
-    public void when_addFavoriteCardNotPresent_thenThrowException() {
-
-        when(mockAccountRepository.findById(1)).thenReturn(Optional.empty());
-        when(mockCardRepository.findById(1)).thenReturn(Optional.empty());
-
-        CardFavoriteDTO dto = new CardFavoriteDTO(1, 1, true);
-        sut.addFavoriteCard(dto);
-    }
-
-    @Test
-    public void when_addFavoriteCardExists_thenreturnTrue() {
-        AccountEntity accountEntity = new AccountEntity();
-        accountEntity.setId(1);
-        accountEntity.setUsername("username");
-        accountEntity.setPassword("password");
-        Set<AccountCardEntity> accountCardEntitySet = new HashSet<>();
-        AccountCardEntity cardEntityAccount = new AccountCardEntity();
-        CardEntity cardEntity = new CardEntity();
-        cardEntity.setQuestion("question");
-        cardEntity.setAnswer("answer");
-        cardEntity.setId(1);
-        cardEntityAccount.setCardEntity(cardEntity);
-        cardEntityAccount.setAccountEntity(accountEntity);
-
-        accountCardEntitySet.add(cardEntityAccount);
-
-        accountEntity.setAccountCardEntities(accountCardEntitySet);
-
-
-        when(mockAccountRepository.findById(1)).thenReturn(Optional.of(accountEntity));
-        when(mockCardRepository.findById(1)).thenReturn(Optional.of(cardEntity));
-
-        CardFavoriteDTO dto = new CardFavoriteDTO(1, 1, true);
-
-        assertTrue(sut.addFavoriteCard(dto));
-    }
 }
